rotating apikey validity is now in seconds
- settings.TASTYPIE_APIKEY_DURATION is now in seconds by default
- use '<n>s|h|d|w|y' for seconds, hours, days, weeks, years
