+++ Makefile
@@ -1,0 +1,3 @@
+test:
+	unset DJANGO_SETTINGS_MODULE; python manage.py test tastypiex
+
